Remove toast-related components and dependencies; refactor toast usage in the application
