--- ExpenseVault.Server/package.json
@@ -1,8 +1,0 @@
-{
-	"devDependencies": {
-		"vite-plugin-svgr": "^4.3.0"
-	},
-	"dependencies": {
-		"react-toastify": "^11.0.5"
-	}
-}
