+++ expensevault.client/src/index.css
@@ -2,6 +2,8 @@
 @import "tailwindcss";
 @plugin "daisyui";
 
-html {
-  font-family: "Poppins", "sans-serif";
+@layer base {
+  html {
+    font-family: "Poppins", sans-serif;
+  }
 }
