--- expensevault.client/src/shared/components/feedback/toast/toast-context.tsx
@@ -1,10 +1,0 @@
-import { createContext } from 'react';
-import { ToastPosition } from 'react-toastify';
-
-type ToastContextType = {
-  position: ToastPosition;
-};
-
-export const ToastContext = createContext<ToastContextType | undefined>(
-  undefined,
-);
